requirements.txt
asgiref==3.6.0
Django==4.1.7
django-appconf==1.0.5
django-cleanup==6.0.0
django-honeypot==1.0.3
django-imagekit==4.1.0
django-tinymce==3.6.1
gunicorn==20.1.0
mysqlclient==2.1.1
pilkit==2.0
Pillow==9.3.0
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
python-decouple==3.6
python-dotenv==0.21.0
six==1.16.0
solid-i18n==1.4.2
sqlparse==0.4.3
whitenoise==6.2.0
